Remove MVP and update tests
