+++ src/applications/debt-letters/components/NeedHelp.jsx
@@ -1,37 +1,58 @@
 import React from 'react';
 
 const NeedHelp = () => (
-  <div className="vads-u-font-family--sans vads-u-margin-top--4">
-    <h3 className="right-heading vads-u-margin-top--0">Need help?</h3>
-    <p>
-      Call the Debt Management Center between 6:30 a.m. and 6:00 p.m. CT using
-      the numbers below:
+  <div className="vads-u-font-family--sans">
+    <h3
+      id="howDoIGetHelp"
+      className="vads-u-margin-top--4 vads-u-margin-bottom--0 vads-u-font-size--h2"
+    >
+      How do I get financial help?
+    </h3>
+    <p className="vads-u-margin-top--2">
+      If you need financial help, you can request:
     </p>
-    <p className="vads-u-margin-top--0">
-      Toll-free:{' '}
+    <ul>
+      <li>
+        An extended monthly payment plan, <strong>or</strong>
+      </li>
+      <li>
+        A compromise (ask us to accept a lower amount of money as full payment
+        of the debt), <strong>or</strong>
+      </li>
+      <li>A waiver (ask us to stop collection on the debt)</li>
+    </ul>
+    <p>
+      You may be required to submit a financial status report. Call the DMC at{' '}
       <a href="tel: 800-827-0648" aria-label="800. 8 2 7. 0648.">
         800-827-0648
-      </a>
-      {'.'}
-    </p>
-    <p className="vads-u-margin-top--0">
-      International:{' '}
+      </a>{' '}
+      between 6:30 a.m. and 6:00 p.m. CT discuss your options and next steps.
+      For international callers, use{' '}
       <a href="tel: 612-713-6415" aria-label="612. 7 1 3. 6415.">
         612-713-6415
       </a>
       {'.'}
     </p>
-    <h5 className="vads-u-margin-top--1 ">Request financial assistance</h5>
-    <p className="vads-u-margin-top--0">
-      If you need financial help, you may request an extended monthly payment
-      plan, compromise, or a waiver. Call the DMC to discuss your options and
-      next steps.
-    </p>
-    <h5 className="vads-u-margin-top--1">Dispute a debt</h5>
-    {/* ToDo: add debt dispute link */}
-    <p className="vads-u-margin-top--0">
-      If you think a debt was created in error, you can dispute it. Call the DMC
-      to get more information.
+    <a href="https://www.va.gov/debtman/Financial_Status_Report.asp">
+      Find information about submitting a financial status report
+    </a>
+    <h3
+      id="howDoIDispute"
+      className="vads-u-margin-top--4 vads-u-margin-bottom--0 vads-u-font-size--h2"
+    >
+      How do I dispute a debt?
+    </h3>
+    <p className="vads-u-margin-top--2">
+      If you think a debt was created in error, you can dispute it. Get
+      information about disputing a debt by calling the DMC at{' '}
+      <a href="tel: 800-827-0648" aria-label="800. 8 2 7. 0648.">
+        800-827-0648
+      </a>{' '}
+      between 6:30 a.m. and 6:00 p.m. CST. For international callers, use{' '}
+      <a href="tel: 612-713-6415" aria-label="612. 7 1 3. 6415.">
+        612-713-6415
+      </a>
+      {'.'}
     </p>
   </div>
 );
