+++ src/applications/debt-letters/tests/DebtLettersSummary.unit.spec.js
@@ -4,8 +4,8 @@ import { expect } from 'chai';
 import DebtLettersSummary from '../components/DebtLettersSummary';
 
 describe('DebtLettersSummary', () => {
-  it('renders correct summary component', () => {
-    const fakeStoreV1 = {
+  it('renders summary component', () => {
+    const fakeStore = {
       getState: () => ({
         featureToggles: {
           debtLettersShowLetters: true,
@@ -25,45 +25,10 @@ describe('DebtLettersSummary', () => {
       subscribe: () => {},
       dispatch: () => {},
     };
-    const wrapper = shallow(<DebtLettersSummary store={fakeStoreV1} />);
-    expect(
-      wrapper.dive().find(`Connect(DebtLettersSummaryV1)`).length,
-    ).to.equal(1);
-    expect(
-      wrapper.dive().find(`Connect(DebtLettersSummaryV2)`).length,
-    ).to.equal(0);
-    wrapper.unmount();
-  });
-
-  it('renders correct summary component V2', () => {
-    const fakeStoreV1 = {
-      getState: () => ({
-        featureToggles: {
-          debtLettersShowLetters: true,
-          debtLettersShowLettersV2: true,
-        },
-        user: {
-          login: {
-            currentlyLoggedIn: true,
-          },
-        },
-        debtLetters: {
-          isFetching: false,
-          isVBMSError: true,
-          isError: true,
-          debts: [],
-        },
-      }),
-      subscribe: () => {},
-      dispatch: () => {},
-    };
-    const wrapper = shallow(<DebtLettersSummary store={fakeStoreV1} />);
-    expect(
-      wrapper.dive().find(`Connect(DebtLettersSummaryV1)`).length,
-    ).to.equal(0);
-    expect(
-      wrapper.dive().find(`Connect(DebtLettersSummaryV2)`).length,
-    ).to.equal(1);
+    const wrapper = shallow(<DebtLettersSummary store={fakeStore} />);
+    expect(wrapper.dive().find(`Connect(DebtLettersSummary)`).length).to.equal(
+      0,
+    );
     wrapper.unmount();
   });
 });
